refactor(debug): inline C/C++ DAP setup into debug plugin spec
dap_cpp.lua was a separate module used in exactly one place.
Inlined setup_cpp_dap() directly into debug.lua to remove
the unnecessary indirection.
